Split OmegaKokkos.h
